Changed meets_requirements to remove redundant checks and refactored to move menu from Content to it's own folder under Structure.
